Add prompt on filechange option

+++ src/fsRenamer.js
@@ -1,7 +1,11 @@
 import fs from 'fs-promise';
 import uuid from 'uuid';
 import _ from 'lodash';
-import {FORCE, BACKUP} from './flags';
+import keypressPrompt from 'keypress-prompt';
+import {FORCE, BACKUP, INTERACTIVE} from './flags';
+
+// assign to variable to enable dependency injection
+let prompt = keypressPrompt.prompt;
 
 let createTempFilename = (filename) =>
     `/tmp/${filename}-${uuid.v4()}`;
@@ -31,10 +35,25 @@ let moveToTemp = async (mediated) => {
     }
 }
 
+let shouldRename = async (src, dest, options) => {
+    let result = true;
+    if (options[INTERACTIVE]) {
+        const confirm = 'y';
+        let response = await prompt(`Renaming ${src} to ${dest}, are you sure?`, [confirm, 'n']);
+        result = (response === confirm);
+    }
+    return result;
+}
+
 let moveToDest = async (mediated, options) => {
     for (const [ src, tmp, dest ] of mediated) {
         try {
-            await fs.move(tmp, dest, { clobber: options[FORCE] });
+            const rename = await shouldRename(src, dest, options);
+            if (rename) {
+                await fs.move(tmp, dest, { clobber: options[FORCE] });
+            } else {
+                await fs.move(tmp, src);
+            }
         } catch (err) {
             if (err.code === 'EEXIST') {
                 await fs.move(tmp, src);
@@ -53,6 +72,11 @@ let fsRenamer = async (pairs, options = {}) => {
     }
     await moveToTemp(mediated);
     await moveToDest(mediated, options);
+}
+
+// dependency injection for testing purposes:
+fsRenamer.__inject = (injectedPrompt) => {
+    prompt = injectedPrompt || prompt
 }
 
 export default fsRenamer;
